+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] 
 
+### Modified
+
+- [TD-3878] Concepts only show links to structures for which user has view_data_structure permission
+
 ### Added
 
 - [TD-3230] Taxonomy aggregations with enriched information